Refactor CommandBuilderFactoryGenerator
Clarify the generator pipeline and the interaction between each step,
improve caching efficiency where possible, and split the code up into
separate files for readability. Simplify the code generation logic for
declaratively-configured commands.
