--- XO.Console.Cli.Analyzers/Model/CommandDeclaration.cs
@@ -1,7 +1,0 @@
-﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
-
-namespace XO.Console.Cli.Model;
-
-internal sealed record CommandDeclaration(
-    ClassDeclarationSyntax SyntaxNode,
-    string ParametersType);
